requirements.txt
aiogram
aiohttp
beautifulsoup4
lxml
sqlalchemy
aiosqlite
python-dotenv
apscheduler
